Move files and delete comment
