+++ src/components/App.js
@@ -1,5 +1,5 @@
-import BooksList from '../components/BooksList';
-import BooksForm from '../components/BooksForm';
+import BooksList from '../containers/BooksList';
+import BooksForm from '../containers/BooksForm';
 
 const App = () => (
   <div>
