+++ src/redux/actions/index.js
@@ -1,5 +1,4 @@
 import { CREATE_BOOK } from './actionTypes';
-// const REMOVE_BOOK = 'REMOVE_BOOK';
 
 const createBook = (book) => ({
   type: CREATE_BOOK,
